move loading system intake into system intake
